+++ assets/javascripts/kitten/components/steppers/next/stepper/index.js
@@ -5,11 +5,7 @@ import COLORS from '../../../../constants/colors-config'
 import { ScreenConfig } from '../../../../constants/screen-config'
 import classNames from 'classnames'
 import { Text } from '../../../typography/text'
-import {
-  StepperIconDefault,
-  StepperIconInProgress,
-  StepperIconValidated,
-} from '../../../../components/steppers/stepper-icon'
+import { StepperIcon } from '../../../../components/atoms/stepper-icon'
 import PropTypes from 'prop-types'
 
 const Wrapper = styled.div`
@@ -38,14 +34,21 @@ const List = styled.ul`
 const ItemWrapper = styled.li`
   display: flex;
   align-items: center;
+
   :not(:last-child) {
-    padding-right: ${pxToRem(50)};
+    margin-right: ${pxToRem(25)};
   }
-  @media (max-width: ${pxToRem(ScreenConfig.S.max)}) {
+
+  @media (min-width: ${pxToRem(ScreenConfig.M.min)}) {
     :not(:last-child) {
-      padding-right: ${pxToRem(25)};
+      margin-right: ${pxToRem(50)};
     }
   }
+
+  .k-Stepper__icon {
+    margin-right: ${pxToRem(10)};
+  }
+
   ${props =>
     props.state === 'progress' &&
     css`
@@ -72,22 +75,11 @@ const StepperText = styled(Text)`
   }
 `
 
-const StepperIcon = ({ state }) => {
-  switch (state) {
-    case 'progress':
-      return <StepperIconInProgress />
-    case 'validated':
-      return <StepperIconValidated />
-    default:
-      return <StepperIconDefault />
-  }
-}
-
 export const StepperItem = ({ children, state, ...props }) => {
   return (
     <ItemWrapper state={state} {...props}>
       <StepperText weight="regular" size="tiny" tabIndex="0">
-        <StepperIcon state={state} />
+        <StepperIcon className="k-Stepper__icon" state={state} />
         {children}
       </StepperText>
     </ItemWrapper>
@@ -127,7 +119,7 @@ export const StepperLink = ({
         rel={external ? 'nofollow noopener noreferrer' : ''}
         {...linkProps}
       >
-        <StepperIcon state={state} />
+        <StepperIcon className="k-Stepper__icon" state={state} />
         {children}
       </StepperText>
     </ItemWrapper>