+++ assets/javascripts/kitten/index.js
@@ -13,6 +13,12 @@ export { Badge } from './components/atoms/badge'
 export { IconBadge } from './components/atoms/icon-badge'
 export { Loader } from './components/atoms/loader'
 export { LoaderWithParagraph } from './components/atoms/loader-with-paragraph'
+export {
+  StepperIcon,
+  StepperIconDefault,
+  StepperIconInProgress,
+  StepperIconValidated,
+} from './components/atoms/stepper-icon'
 export { Tag } from './components/atoms/tag'
 
 // Avatar
@@ -196,11 +202,6 @@ export { StatusWithBullet } from './components/status-with-bullet'
 
 // Steppers
 export { VerticalStepper } from './components/steppers/vertical-stepper'
-export {
-  StepperIconDefault,
-  StepperIconInProgress,
-  StepperIconValidated,
-} from './components/steppers/stepper-icon'
 export { Stepper } from './components/steppers/stepper'
 
 // Tables