+++ CHANGELOG.md
@@ -4,9 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Breaking change: convert `Stepper` to CSS-in-JS.
+- Breaking change: remove `k-Stepper` mixin and `k-Stepper__icon` style.
+- Breaking change: convert and move `StepperIcon*` components to `atoms`.
+- Feature: Add `StepperIcon` component with `state` prop.
 - Breaking change: convert `Line` to CSS-in-JS.
 - Breaking change: remove `k-Line` mixin.
 - Feature: Add `useFlexGapCheck` hook for `gap` feature detection.
+
 
 ## [3.0.0-beta.5] - 2021-02-25